Add the hypergen feature.
This feature is not something I'd really recommend that anyone else
use because I'm not really confident with it. Hence I'm leaving it
undocumented. I personally feel like the code for this feature is
pretty messy and hackish; I'm not really proud of it (I feel
rather ashamed about it actually). But it works and it's useful so
might as well commit it.

One thing this made me realize is that in the long run all of the
stations' data should be loaded into, and maintained as, one large
dictionary in memory instead of having this silly list nonsense.
Just because the PSD has ugly and inconsistent formatting shouldn't
mean that I have to let their messiness screw with my code. For
example for industry, I can have two elements in each station dict
entry - one would be a "clean" version of industry and one would be a
messy un-stripped version.

But since this probably won't be used after this year I don't want
to invest in the time to refactor it. Even if it is used later, I'll
let someone else do it. As things stand I have little to nothing to
gain from it (mainly because this is a niche tool used only by a few
people that won't need too many upgrades going forward).